+++ classes/Task/AbstractTask.php
@@ -25,7 +25,7 @@
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License 3.0 (AFL-3.0)
  */
 
-namespace PrestaShop\Module\AutoUpgrade\TaskRunner;
+namespace PrestaShop\Module\AutoUpgrade\Task;
 
 use Exception;
 use PrestaShop\Module\AutoUpgrade\AjaxResponse;
