+++ classes/Task/Rollback/NoRollbackFound.php
@@ -25,42 +25,18 @@
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License 3.0 (AFL-3.0)
  */
 
-namespace PrestaShop\Module\AutoUpgrade\TaskRunner\Rollback;
+namespace PrestaShop\Module\AutoUpgrade\Task\Rollback;
 
-use PrestaShop\Module\AutoUpgrade\TaskRunner\ChainedTasks;
+use PrestaShop\Module\AutoUpgrade\Task\AbstractTask;
+use PrestaShop\Module\AutoUpgrade\Task\ExitCode;
 
-/**
- * Execute the whole upgrade process in a single request.
- */
-class AllRollbackTasks extends ChainedTasks
+class NoRollbackFound extends AbstractTask
 {
-    const initialTask = 'rollback';
-
-    /**
-     * @var string
-     */
-    protected $step = self::initialTask;
-
-    /**
-     * Customize the execution context with several options
-     * > action: Replace the initial step to run
-     * > channel: Makes a specific upgrade (minor, major etc.)
-     * > data: Loads an encoded array of data coming from another request.
-     *
-     * @param array<string, string> $options
-     */
-    public function setOptions(array $options): void
+    public function run(): int
     {
-        if (!empty($options['backup'])) {
-            $this->container->getState()->setRestoreName($options['backup']);
-        }
-    }
+        $this->logger->info($this->translator->trans('Nothing to restore'));
+        $this->next = 'rollbackComplete';
 
-    /**
-     * Set default config on first run.
-     */
-    public function init(): void
-    {
-        // Do nothing
+        return ExitCode::SUCCESS;
     }
 }
