+++ classes/Task/Upgrade/AllUpgradeTasks.php
@@ -25,11 +25,11 @@
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License 3.0 (AFL-3.0)
  */
 
-namespace PrestaShop\Module\AutoUpgrade\TaskRunner\Upgrade;
+namespace PrestaShop\Module\AutoUpgrade\Task\Upgrade;
 
 use Exception;
 use PrestaShop\Module\AutoUpgrade\AjaxResponse;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\ChainedTasks;
+use PrestaShop\Module\AutoUpgrade\Task\ChainedTasks;
 use PrestaShop\Module\AutoUpgrade\UpgradeContainer;
 
 /**
