+++ classes/Task/Upgrade/CleanDatabase.php
@@ -25,17 +25,29 @@
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License 3.0 (AFL-3.0)
  */
 
-namespace PrestaShop\Module\AutoUpgrade\TaskRunner\Rollback;
+namespace PrestaShop\Module\AutoUpgrade\Task\Upgrade;
 
-use PrestaShop\Module\AutoUpgrade\TaskRunner\AbstractTask;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\ExitCode;
+use PrestaShop\Module\AutoUpgrade\Task\AbstractTask;
+use PrestaShop\Module\AutoUpgrade\Task\ExitCode;
 
-class NoRollbackFound extends AbstractTask
+/**
+ * Clean the database from unwanted entries.
+ */
+class CleanDatabase extends AbstractTask
 {
     public function run(): int
     {
-        $this->logger->info($this->translator->trans('Nothing to restore'));
-        $this->next = 'rollbackComplete';
+        // Clean tabs order
+        foreach ($this->container->getDb()->ExecuteS('SELECT DISTINCT id_parent FROM ' . _DB_PREFIX_ . 'tab') as $parent) {
+            $i = 1;
+            foreach ($this->container->getDb()->ExecuteS('SELECT id_tab FROM ' . _DB_PREFIX_ . 'tab WHERE id_parent = ' . (int) $parent['id_parent'] . ' ORDER BY IF(class_name IN ("AdminHome", "AdminDashboard"), 1, 2), position ASC') as $child) {
+                $this->container->getDb()->Execute('UPDATE ' . _DB_PREFIX_ . 'tab SET position = ' . (int) ($i++) . ' WHERE id_tab = ' . (int) $child['id_tab'] . ' AND id_parent = ' . (int) $parent['id_parent']);
+            }
+        }
+
+        $this->status = 'ok';
+        $this->next = 'upgradeComplete';
+        $this->logger->info($this->translator->trans('The database has been cleaned.'));
 
         return ExitCode::SUCCESS;
     }
