+++ classes/UpgradeTools/TaskRepository.php
@@ -27,27 +27,27 @@
 
 namespace PrestaShop\Module\AutoUpgrade\UpgradeTools;
 
-use PrestaShop\Module\AutoUpgrade\TaskRunner\AbstractTask;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Miscellaneous\CheckFilesVersion;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Miscellaneous\CompareReleases;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Miscellaneous\GetChannelInfo;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Miscellaneous\UpdateConfig;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\NullTask;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Rollback\NoRollbackFound;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Rollback\RestoreDb;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Rollback\RestoreFiles;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Rollback\Rollback;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Rollback\RollbackComplete;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Upgrade\BackupDb;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Upgrade\BackupFiles;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Upgrade\CleanDatabase;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Upgrade\Download;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Upgrade\Unzip;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Upgrade\UpgradeComplete;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Upgrade\UpgradeDb;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Upgrade\UpgradeFiles;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Upgrade\UpgradeModules;
-use PrestaShop\Module\AutoUpgrade\TaskRunner\Upgrade\UpgradeNow;
+use PrestaShop\Module\AutoUpgrade\Task\AbstractTask;
+use PrestaShop\Module\AutoUpgrade\Task\Miscellaneous\CheckFilesVersion;
+use PrestaShop\Module\AutoUpgrade\Task\Miscellaneous\CompareReleases;
+use PrestaShop\Module\AutoUpgrade\Task\Miscellaneous\GetChannelInfo;
+use PrestaShop\Module\AutoUpgrade\Task\Miscellaneous\UpdateConfig;
+use PrestaShop\Module\AutoUpgrade\Task\NullTask;
+use PrestaShop\Module\AutoUpgrade\Task\Rollback\NoRollbackFound;
+use PrestaShop\Module\AutoUpgrade\Task\Rollback\RestoreDb;
+use PrestaShop\Module\AutoUpgrade\Task\Rollback\RestoreFiles;
+use PrestaShop\Module\AutoUpgrade\Task\Rollback\Rollback;
+use PrestaShop\Module\AutoUpgrade\Task\Rollback\RollbackComplete;
+use PrestaShop\Module\AutoUpgrade\Task\Upgrade\BackupDb;
+use PrestaShop\Module\AutoUpgrade\Task\Upgrade\BackupFiles;
+use PrestaShop\Module\AutoUpgrade\Task\Upgrade\CleanDatabase;
+use PrestaShop\Module\AutoUpgrade\Task\Upgrade\Download;
+use PrestaShop\Module\AutoUpgrade\Task\Upgrade\Unzip;
+use PrestaShop\Module\AutoUpgrade\Task\Upgrade\UpgradeComplete;
+use PrestaShop\Module\AutoUpgrade\Task\Upgrade\UpgradeDb;
+use PrestaShop\Module\AutoUpgrade\Task\Upgrade\UpgradeFiles;
+use PrestaShop\Module\AutoUpgrade\Task\Upgrade\UpgradeModules;
+use PrestaShop\Module\AutoUpgrade\Task\Upgrade\UpgradeNow;
 use PrestaShop\Module\AutoUpgrade\UpgradeContainer;
 
 class TaskRepository
