+++ cli-rollback.php
@@ -33,7 +33,7 @@ require_once realpath(dirname(__FILE__) . '/../../modules/autoupgrade') . '/ajax
 $container = autoupgrade_init_container(dirname(__FILE__));
 
 $container->setLogger(new PrestaShop\Module\AutoUpgrade\Log\StreamedLogger());
-$controller = new \PrestaShop\Module\AutoUpgrade\TaskRunner\Rollback\AllRollbackTasks($container);
+$controller = new \PrestaShop\Module\AutoUpgrade\Task\Rollback\AllRollbackTasks($container);
 $controller->setOptions(getopt('', ['backup::']));
 $controller->init();
 exit($controller->run());
