+++ tests/fixtures/listOfFiles-ShopExample.json
@@ -6195,35 +6195,35 @@
     "\/modules\/autoupgrade\/classes\/Parameters\/index.php",
     "\/modules\/autoupgrade\/classes\/PrestashopConfiguration.php",
     "\/modules\/autoupgrade\/classes\/State.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/AbstractTask.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/ChainedTasks.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Miscellaneous\/CheckFilesVersion.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Miscellaneous\/CompareReleases.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Miscellaneous\/GetChannelInfo.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Miscellaneous\/UpdateConfig.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Miscellaneous\/index.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/NullTask.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Rollback\/AllRollbackTasks.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Rollback\/NoRollbackFound.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Rollback\/RestoreDb.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Rollback\/RestoreFiles.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Rollback\/Rollback.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Rollback\/RollbackComplete.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Rollback\/index.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/AllUpgradeTasks.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/BackupDb.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/BackupFiles.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/CleanDatabase.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/Download.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/RemoveSamples.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/Unzip.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/UpgradeComplete.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/UpgradeDb.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/UpgradeFiles.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/UpgradeModules.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/UpgradeNow.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/Upgrade\/index.php",
-    "\/modules\/autoupgrade\/classes\/TaskRunner\/index.php",
+    "\/modules\/autoupgrade\/classes\/Task\/AbstractTask.php",
+    "\/modules\/autoupgrade\/classes\/Task\/ChainedTasks.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Miscellaneous\/CheckFilesVersion.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Miscellaneous\/CompareReleases.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Miscellaneous\/GetChannelInfo.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Miscellaneous\/UpdateConfig.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Miscellaneous\/index.php",
+    "\/modules\/autoupgrade\/classes\/Task\/NullTask.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Rollback\/AllRollbackTasks.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Rollback\/NoRollbackFound.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Rollback\/RestoreDb.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Rollback\/RestoreFiles.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Rollback\/Rollback.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Rollback\/RollbackComplete.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Rollback\/index.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/AllUpgradeTasks.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/BackupDb.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/BackupFiles.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/CleanDatabase.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/Download.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/RemoveSamples.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/Unzip.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/UpgradeComplete.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/UpgradeDb.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/UpgradeFiles.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/UpgradeModules.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/UpgradeNow.php",
+    "\/modules\/autoupgrade\/classes\/Task\/Upgrade\/index.php",
+    "\/modules\/autoupgrade\/classes\/Task\/index.php",
     "\/modules\/autoupgrade\/classes\/Tools14.php",
     "\/modules\/autoupgrade\/classes\/Twig\/Block\/ChannelInfoBlock.php",
     "\/modules\/autoupgrade\/classes\/Twig\/Block\/RollbackForm.php",
